requirements.txt
Flask
gunicorn
tensorflow-cpu
opencv-python-headless
numpy
werkzeug
